pyproject.toml
[tool.poetry]
name = "local-intelligence-hub"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
Django = "4.2.10"
django-bootstrap5 = "^22.1"
django-environ = "^0.9.0"
psycopg2 = "^2.9.3"
django-libsass = "^0.9"
requests-cache = "^0.9.6"
Pillow = "^10.2.0"
python-magic = "^0.4.27"
tqdm = "^4.64.1"
openpyxl = "^3.0.10"
mysoc-dataset = "^0.3.0"
django-jsonform = "^2.15.0"
lxml = "^4.9.2"
beautifulsoup4 = "^4.11.1"
django-polymorphic = "^3.1.0"
pyairtable = "^2.3.0.post1"
procrastinate = {version = "2.0.0b4", extras = ["django"]}
psycopg = "^3.1.18"
aiopg = "^1.4.0"
strawberry-graphql-django = "^0.32.1"
strawberry-django-auth = "^0.376.5"
django-cors-headers = "^4.3.1"
python-benedict = "^0.33.2"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
setuptools = "^69.1.1"
uvicorn = "^0.27.1"
strawberry-graphql = {extras = ["asgi"], version = "^0.220.0"}
pandas = "^2.2.1"
django-choices-field = "^2.3.0"
httpx = "^0.27.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
coverage = "^6.5.0"
flake8 = "^6.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
default_section = "THIRDPARTY"
known_first_party = "hub"
known_django = "django"
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
extend_skip = ["migrations"]
